+++ NEWS.md
@@ -1,1 +1,13 @@
 # Virt Tuner News
+
+## Version 0.0.1
+
+* Single VM per host template
+
+## Version 0.0.2
+
+* Fix NUMA cells memory computation for 'single' template
+
+## Version 0.0.3
+
+* Fix the virtual CPUs ordering to have sibling with consecutive IDs.